+++ changelog.md
@@ -5,6 +5,7 @@
 - Added ``%zessentials_count_repair_all%`` placeholder, counting the items to be repaired
 - Added ``/repairall [<player>]``
 - Fixed commands that could not be used from the console
+- Added ``/tpahere <player>``
 
 # 1.0.1.5